Re #1474: merged r4100-4107
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@4108 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjmedia/src/pjmedia/stream.c
@@ -961,7 +961,7 @@ static pj_status_t send_rtcp(pjmedia_stream *stream,
 	    PJ_PERROR(4,(stream->port.info.name.ptr, status,
         			     "Error generating RTCP SDES"));
 	} else {
-	    len += sdes_len;
+	    len += (int)sdes_len;
 	}
     }
 
@@ -1019,7 +1019,7 @@ static pj_status_t send_rtcp(pjmedia_stream *stream,
 	    PJ_PERROR(4,(stream->port.info.name.ptr, status,
         			     "Error generating RTCP BYE"));
 	} else {
-	    len += bye_len;
+	    len += (int)bye_len;
 	}
     }
